+++ src/supermodels/__init__.py
@@ -5,7 +5,7 @@ supermodels
 ...
 """
 
-__version__ = "0.1.0"
+__version__ = "0.1.12"
 __author__ = "Joel Yisrael"
 __email__ = "schizoprada@gmail.com"
 __license__ = "MIT"