+++ src/supermodels/core/hints.py
@@ -9,3 +9,5 @@ if t.TYPE_CHECKING:
     from supermodels.core.bases.manager import BaseManager
 
 MetaModelRegistry = t.Dict[t.Type[t.Any], t.Type['BaseManager']]
+
+ManagerInstanceRegistry = t.Dict[t.Type[t.Any], 'BaseManager']