+++ tests/fixtures/__init__.py
@@ -1,0 +1,1 @@
+# ~/supermodels/tests/fixtures/__init__.py